requirements.txt
alembic==1.7.7
amqp==5.0.9
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.2
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
kombu==5.2.3
Mako==1.2.0
MarkupSafe==2.0.1
orjson==3.6.4
packaging==21.3
passlib==1.7.4
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.7
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
redis==4.2.0
requests==2.26.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
wrapt==1.14.0
